+++ b2sdk/v1/bucket.py
@@ -22,6 +22,7 @@ from b2sdk.utils import validate_b2_file_name
 # and to retain old style download_file_by_id signature (allowing for the new one as well)
 # and to retain old style get_file_info_by_name return type
 # and to to adjust to old style B2Api.get_file_info return type
+# and to retain old style update return type
 class Bucket(v2.Bucket):
     FILE_VERSION_FACTORY = staticmethod(FileVersionInfoFactory)
 
@@ -196,6 +197,40 @@ class Bucket(v2.Bucket):
         :param str file_id: the id of the file who's info will be retrieved.
         """
         return self.api.file_version_factory.from_api_response(self.api.get_file_info(file_id))
+
+    def update(
+        self,
+        bucket_type: Optional[str] = None,
+        bucket_info: Optional[dict] = None,
+        cors_rules: Optional[dict] = None,
+        lifecycle_rules: Optional[dict] = None,
+        if_revision_is: Optional[int] = None,
+        default_server_side_encryption: Optional[v2.EncryptionSetting] = None,
+        default_retention: Optional[v2.BucketRetentionSetting] = None,
+    ):
+        """
+        Update various bucket parameters.
+
+        :param str bucket_type: a bucket type
+        :param dict bucket_info: an info to store with a bucket
+        :param dict cors_rules: CORS rules to store with a bucket
+        :param dict lifecycle_rules: lifecycle rules to store with a bucket
+        :param int if_revision_is: revision number, update the info **only if** *revision* equals to *if_revision_is*
+        :param b2sdk.v1.EncryptionSetting default_server_side_encryption: default server side encryption settings (``None`` if unknown)
+        :param b2sdk.v1.BucketRetentionSetting default_retention: bucket default retention setting
+        """
+        account_id = self.api.account_info.get_account_id()
+        return self.api.session.update_bucket(
+            account_id,
+            self.id_,
+            bucket_type=bucket_type,
+            bucket_info=bucket_info,
+            cors_rules=cors_rules,
+            lifecycle_rules=lifecycle_rules,
+            if_revision_is=if_revision_is,
+            default_server_side_encryption=default_server_side_encryption,
+            default_retention=default_retention,
+        )
 
 
 def download_file_and_return_info_dict(